requirements.txt
openai
pyyaml
pyside6
markdown